fix : fixed area dimensions also based on width and height

+++ apps/api/src/modules/admin/registry.ts
@@ -136,8 +136,8 @@ export const TABLES: TableDef[] = [
     fields: [
       // manufacturerId is the normalised manufacturer FK (U0). The generic CRUD has no dedicated
       // `ref` field type, so it is exposed as an int FK (manufacturers.id).
-      f('vendor'), f('manufacturerId', 'int'), f('model', 'string', true), f('serviceCategory'), f('moduleWMm', 'int'),
-      f('moduleHMm', 'int'), f('minCabinetWMm', 'int'), f('minCabinetHMm', 'int'), f('cabinetDepthMm', 'int'),
+      f('vendor'), f('manufacturerId', 'int'), f('model', 'string', true), f('serviceCategory'), f('moduleWMm', 'decimal'),
+      f('moduleHMm', 'decimal'), f('minCabinetWMm', 'decimal'), f('minCabinetHMm', 'decimal'), f('cabinetDepthMm', 'int'),
       f('cabinetType'), f('pixelPitchH', 'decimal'), f('pixelPitchV', 'decimal'), f('brightnessNits', 'int'),
       // W0: indoor/outdoor suitability (nullable enum). Null → config falls back to a brightness heuristic.
       f('environment', 'enum', false, ['indoor', 'outdoor']),
